Move `AppBuilder` and a few type aliases into a separate file.
